Created the Odin Project: Landing Page
Sorry for not committing during the process, major mistake on my end.
Created the page to include pictures and things about Stitch.
This is the Stitch from the movie and show "Lilo & Stitch".
Included links and credit to images found on the internet in README.
Used a lot of flexbox techniques and features.